fix deleting image button

+++ src/firebase/storageFunctions.js
@@ -19,6 +19,8 @@ export const uploadFile = (collectionImageInputId, uid, collectionId) => {
 }
 
 export const deleteFile = imageUrl => {
-  const httpsRef = storage.refFromURL(imageUrl)
-  return httpsRef.delete().catch(error => console.log(error))
+  if (imageUrl.includes('firebasestorage')) {
+    const httpsRef = storage.refFromURL(imageUrl)
+    return httpsRef.delete().catch(error => console.log(error))
+  } else return null
 }
